+++ src/aac/libfaad2/common.h
@@ -228,7 +228,7 @@ typedef float float32_t;
 # else
 /* we need these... */
 typedef unsigned long long uint64_t;
-typedef unsigned long uint32_t;
+typedef unsigned int uint32_t;
 typedef unsigned short uint16_t;
 typedef unsigned char uint8_t;
 typedef long long int64_t;
